backend/requirements.txt
# SonicGuard — Plagiarism Detection Engine (DSP + API)
# Zero dependências de IA/ML (sem torch, transformers, ou modelos neurais)

# Motor DSP
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
yt-dlp>=2024.0
soundfile>=0.12.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# Legal Module (Gemini Flash)
google-generativeai>=0.5.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# Database & Authentication (Phase 4)
sqlalchemy>=2.0.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0


